Validerte gjennom Hibernate constraint
