+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vilkårsvurdering/VilkårResultat.kt
@@ -171,24 +171,6 @@ class VilkårResultat(
     fun erAvslagUtenPeriode() = this.erEksplisittAvslagPåSøknad == true && this.periodeFom == null && this.periodeTom == null
     fun harFremtidigTom() = this.periodeTom == null || this.periodeTom!!.isAfter(LocalDate.now().sisteDagIMåned())
 
-    fun validerOpsjoner(): List<String> {
-        val listeAvFeil = mutableListOf<String>()
-
-        if (vilkårType != Vilkår.BOSATT_I_RIKET && erMedlemskapVurdert) {
-            listeAvFeil.add("Vilkår ${vilkårType.spesifikasjon.beskrivelse} skal ikke ha satt feltet 'Medlemskap vurdert'. Dette skal kun kunne settes for ${Vilkår.BOSATT_I_RIKET}")
-        }
-
-        if (vilkårType != Vilkår.BOR_MED_SØKER && erDeltBosted) {
-            listeAvFeil.add("Vilkår ${vilkårType.spesifikasjon.beskrivelse} skal ikke ha satt feltet 'Delt bosted'. Dette skal kun kunne settes for ${Vilkår.BOR_MED_SØKER}")
-        }
-
-        if (vilkårType != Vilkår.BOSATT_I_RIKET && erSkjønnsmessigVurdert) {
-            listeAvFeil.add("Vilkår ${vilkårType.spesifikasjon.beskrivelse} skal ikke ha satt feltet 'Vurdering annet grunnlag'. Dette skal kun kunne settes for ${Vilkår.BOSATT_I_RIKET}")
-        }
-
-        return listeAvFeil
-    }
-
     val vedtaksperiodeFom
         get() =
             if (this.vilkårType == Vilkår.UNDER_18_ÅR) this.periodeFom?.førsteDagIInneværendeMåned()
